chore: last fixes to finish refactor
