+++ README.md
@@ -1,47 +1,3 @@
-# New captcha update:
-
-### Ta oficialmente terminado!
-### It is done!
-  
-
-## Faz a boa pra nois
-## Do the good
-
-### Smart Chain Wallet(BUSD/BNB/BCOIN):
-#### 0xbd06182D8360FB7AC1B05e871e56c76372510dDf
-
-### PIX:
-  6ce2b79e-97be-4881-a882-81902e29d7ce
-![pix](https://github.com/mpcabete/bombcrypto-bot/raw/main/readme-images/pix.jpeg)
-
-  
-# Sobre:
-Este bot tem o seu código aberto, de forma que qualquer pessoa pode vê-lo, fazer uma fork, ou updates.
-
-Desenvolvi esse bot inicialmente para o meu uso pessoal. Eu decidi publica-lo
-aqui para ajudar o pessoal e com a esperança de ganhar um trocadinho com
-doações.  Com o tempo mais e mais pessoas foram abrindo issues, pedindo ajuda,
-e dando sugestões. Eu tento responder todo mundo, mas ultimamente tem sido
-difícil acompanhar a demanda. 
-
-Eu gostaria de manter este bot grátis e com o código aberto. Para que isso
-seja possível eu estou criando algumas metas de doação para que o bot possa
-ser financiado coletivamente. Atualmente eu atualizarei a barra das metas
-manualmente de forma diária, talvez no futuro eu automatize o processo de
-alguma forma.
-
- 
-
-### Paypal:
-[Donate:](https://www.paypal.com/donate?hosted_button_id=JVYSC6ZYCNQQQ)
-https://www.paypal.com/donate?hosted_button_id=JVYSC6ZYCNQQQ
-or
-mpcabete@protonmail.com
-
-## Aviso:
-
-#### Os desenvolvedores do jogo se pronunciaram e agora o uso de auto clickers e o uso de bots é oficialmente PROIBIDO.  Não me responsabilizo por eventuais penalidades sofridas por quem usar o bot, use por sua própria conta e risco.
-
 # Instalação:
 ### Baixe e instale o Python pelo [site](https://www.python.org/downloads/) ou pela [windows store](https://www.microsoft.com/p/python-37/9nj46sx7x90p?activetab=pivot:overviewtab).
 
@@ -216,11 +172,3 @@ exemplo na pasta “targets” para uma tirada no seu próprio computador:
 
 
 ### Algumas configuraçoes podem ser mudadas no arquivo config.yaml, nao se esqueça de reiniciar o bot caso mude as configuraçoes.
-
-## Curtiu? Dê aquela fortalecida :)
-
-### Wallet:
-#### 0xbd06182D8360FB7AC1B05e871e56c76372510dDf
-### Paypal:
-[Donate](https://www.paypal.com/donate?hosted_button_id=JVYSC6ZYCNQQQ)
-
